Create SignInPage
As Fadhil suggested, it would be more clear if we have a components
folder and a separate folder for pages, since pages can consist
of more components.
