+++ src/styling/style.scss
@@ -10,8 +10,8 @@
 @import '../components/GoogleMap/GoogleMap';
 @import '../components/Loader/Loader';
 @import '../components/Navbar/Navbar';
-@import '../components/PasswordForget/PasswordForget';
 @import '../components/PasswordChange/PasswordChange';
-@import '../components/SignIn/SignIn';
-@import '../components/SignOut/SignOut';
-@import '../components/SignUp/SignUp';
+@import '../components/Authentication/PasswordForget';
+@import '../components/Authentication/SignIn';
+@import '../components/Authentication/SignOut';
+@import '../components/Authentication/SignUp';
